fix: bug in adding game time
