+++ app/chore-chart/UserStatsCard.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import assert from 'assert';
-import React, {useEffect, useState} from 'react';
+import React, { useEffect, useState } from 'react';
 import { useAuth } from '../AuthContext';
 import { useRouter } from 'next/navigation';
 import { StatBox, TimeModifier, PositiveIntegerPicker } from '../StatBox';
-import { update_reward_stars, update_reward_time } from '../Utils';
+import { update_reward_stars, update_reward_time, convert_minutes_to_display_time } from '../Utils';
 import { MONEY_PER_STAR_POINT } from '../Constants';
 
 
@@ -19,23 +19,6 @@ interface UserStatsCardProps {
 }
 
 
-function convert_minutes_to_display_time(minutes: number, show_sign: boolean = true): string
-{
-  const sign = minutes < 0 ? '-' : '+';
-  const absMinutes = Math.abs(minutes);
-  const hours = Math.floor((absMinutes % (24 * 60)) / 60);
-  const mins = absMinutes % 60;
-
-  const hoursString = hours.toString().padStart(2, '0');
-  const minutesString = mins.toString().padStart(2, '0');
-  if (show_sign) {
-    return `${sign}${hoursString}:${minutesString}`.trim();
-  }
-
-  return `${hoursString}:${minutesString}`.trim();
-}
-
-
 const UserStatsCard: React.FC<UserStatsCardProps> = ({
   starPoints,
   gameTime,
@@ -45,11 +28,11 @@ const UserStatsCard: React.FC<UserStatsCardProps> = ({
   pendingPoints,
 }) => {
 
-  const [ money, setMoney ] = useState(starPoints * MONEY_PER_STAR_POINT);
+  const [money, setMoney] = useState(starPoints * MONEY_PER_STAR_POINT);
   const router = useRouter();
   const { active_user } = useAuth();
 
-  assert (active_user !== null, "Active user is null");
+  assert(active_user !== null, "Active user is null");
 
   const pendingGameTimeClass = pendingGameTime >= 0 ? 'font-bold' : 'text-red-500';
   const pendingGameTimeDisplay = convert_minutes_to_display_time(pendingGameTime);
@@ -63,7 +46,7 @@ const UserStatsCard: React.FC<UserStatsCardProps> = ({
     setMoney(starPoints * MONEY_PER_STAR_POINT);
   }, [starPoints]);
 
-  const update_star_points = async (value: string|number, operation: string, amount: string|number) =>   {
+  const update_star_points = async (value: string | number, operation: string, amount: string | number) => {
     const new_stars_str = await update_reward_stars(active_user.id, 'star_points', value, operation, amount);
     const new_stars = parseInt(new_stars_str);
     setMoney(new_stars * MONEY_PER_STAR_POINT);
@@ -71,13 +54,11 @@ const UserStatsCard: React.FC<UserStatsCardProps> = ({
   }
 
   const handleAvatarClick = () => {
-    if(active_user === null || active_user === undefined)
-    {
+    if (active_user === null || active_user === undefined) {
       return;
     }
 
-    if(active_user.is_superuser)
-    {
+    if (active_user.is_superuser) {
       router.push('/dashboard');
     }
   };
@@ -96,49 +77,49 @@ const UserStatsCard: React.FC<UserStatsCardProps> = ({
         </div>
 
         <StatBox icon="🎮"
-                 label="Game"
-                 initialValue={gameTimeDisplay}
-                 defaultValue="00:00"
-                 renderPicker={TimeModifier}
-                 applyOperation={(val, op, amount) => update_reward_time(active_user.id, 'game_time', val, op, amount)}
-                 renderContent={({ value }) => (
-                   <div className="flex flex-col items-center">
-                     <div className="font-bold text-white text-lg">{value}</div>
-                     <div className={`text-xs text-white opacity-80 ${pendingGameTimeClass}`}>{pendingGameTimeDisplay}</div>
-                   </div>)}
+          label="Game"
+          initialValue={gameTimeDisplay}
+          defaultValue="00:00"
+          renderPicker={TimeModifier}
+          applyOperation={(val, op, amount) => update_reward_time(active_user.id, 'game_time', val, op, amount)}
+          renderContent={({ value }) => (
+            <div className="flex flex-col items-center">
+              <div className="font-bold text-white text-lg">{value}</div>
+              <div className={`text-xs text-white opacity-80 ${pendingGameTimeClass}`}>{pendingGameTimeDisplay}</div>
+            </div>)}
         />
 
         <StatBox icon="📺"
-                 label="TV"
-                 initialValue={tvTimeDisplay}
-                 defaultValue="00:00"
-                 renderPicker={TimeModifier}
-                 applyOperation={(val, op, amount) => update_reward_time(active_user.id, 'tv_time', val, op, amount)}
-                 renderContent={({ value }) => (
-                   <div className="flex flex-col items-center">
-                     <div className="font-bold text-white text-lg">{value}</div>
-                     <div className={`text-xs text-white opacity-80 ${pendingTimeClass}`}>{pendingTVTimeDisplay}</div>
-                   </div>)}
+          label="TV"
+          initialValue={tvTimeDisplay}
+          defaultValue="00:00"
+          renderPicker={TimeModifier}
+          applyOperation={(val, op, amount) => update_reward_time(active_user.id, 'tv_time', val, op, amount)}
+          renderContent={({ value }) => (
+            <div className="flex flex-col items-center">
+              <div className="font-bold text-white text-lg">{value}</div>
+              <div className={`text-xs text-white opacity-80 ${pendingTimeClass}`}>{pendingTVTimeDisplay}</div>
+            </div>)}
         />
 
         <StatBox icon="🌟"
-                 label="Stars"
-                 initialValue={starPoints}
-                 defaultValue={0}
-                 renderPicker={PositiveIntegerPicker}
-                 applyOperation={update_star_points}
-                 renderContent={({ value }) => (
-                   <div className="flex flex-col items-center">
-                     <div className="font-bold text-white text-lg">{value}</div>
-                     <div className={`text-xs text-white opacity-80`}>{pendingPoints}</div>
-                   </div>)}
+          label="Stars"
+          initialValue={starPoints}
+          defaultValue={0}
+          renderPicker={PositiveIntegerPicker}
+          applyOperation={update_star_points}
+          renderContent={({ value }) => (
+            <div className="flex flex-col items-center">
+              <div className="font-bold text-white text-lg">{value}</div>
+              <div className={`text-xs text-white opacity-80`}>{pendingPoints}</div>
+            </div>)}
         />
 
-          <div className="bg-white bg-opacity-20 rounded-xl p-2 text-white">
-            <div className="font-medium">💵 Value</div>
-            <div className="text-lg font-bold">${money.toFixed(2)}</div>
-            <span className={`text-xs text-white opacity-80`}>${pendingMoney.toFixed(2)}</span>
-          </div>
+        <div className="bg-white bg-opacity-20 rounded-xl p-2 text-white">
+          <div className="font-medium">💵 Value</div>
+          <div className="text-lg font-bold">${money.toFixed(2)}</div>
+          <span className={`text-xs text-white opacity-80`}>${pendingMoney.toFixed(2)}</span>
+        </div>
       </div>
     </div>
   );
