+++ app/dashboard/DashboardUserStats.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import React, {useEffect, useState} from 'react';
+import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { FamilyMember } from '../types';
 import { StatBox, TimeModifier, PositiveIntegerPicker } from '../StatBox';
-import { update_reward_stars, update_reward_time } from '../Utils';
-import {MONEY_PER_STAR_POINT} from '../Constants';
+import { update_reward_stars, update_reward_time, convert_minutes_to_display_time } from '../Utils';
+import { MONEY_PER_STAR_POINT } from '../Constants';
 import { useAlert } from '../AlertContext';
 
 
@@ -18,23 +18,6 @@ interface DashboardUserStatsProps {
 }
 
 
-function convert_minutes_to_display_time(minutes: number, show_sign: boolean = true): string
-{
-  const sign = minutes < 0 ? '-' : '+';
-  const absMinutes = Math.abs(minutes);
-  const hours = Math.floor((absMinutes % (24 * 60)) / 60);
-  const mins = absMinutes % 60;
-
-  const hoursString = hours.toString().padStart(2, '0');
-  const minutesString = mins.toString().padStart(2, '0');
-  if (show_sign) {
-    return `${sign}${hoursString}:${minutesString}`.trim();
-  }
-
-  return `${hoursString}:${minutesString}`.trim();
-}
-
-
 const DashboardUserStats: React.FC<DashboardUserStatsProps> = ({
   user,
   gameTime,
@@ -43,44 +26,39 @@ const DashboardUserStats: React.FC<DashboardUserStatsProps> = ({
   onWeekEnd
 }) => {
 
-  const [ money, setMoney ] = useState(stars !== undefined ? stars * MONEY_PER_STAR_POINT : 0);
+  const [money, setMoney] = useState(stars !== undefined ? stars * MONEY_PER_STAR_POINT : 0);
   const { showAlert } = useAlert();
   const gameTimeDisplay: string = convert_minutes_to_display_time(gameTime, false);
   const tvTimeDisplay = convert_minutes_to_display_time(tvTime, false);
 
   useEffect(() => {
-    if(stars !== undefined && stars !== null)
-    {
+    if (stars !== undefined && stars !== null) {
       setMoney(stars * MONEY_PER_STAR_POINT);
     }
   }, [stars]);
 
-  const update_star_points = async (value: string|number, operation: string, amount: string|number) =>   {
+  const update_star_points = async (value: string | number, operation: string, amount: string | number) => {
     const new_stars_str = await update_reward_stars(user.id, 'star_points', value, operation, amount);
     const new_stars = parseInt(new_stars_str);
     setMoney(new_stars * MONEY_PER_STAR_POINT);
     return new_stars_str;
   }
 
-  const handle_week_end = async () =>
-  {
+  const handle_week_end = async () => {
     const params = { user_id: user.id };
     const response = await axios.post('/backend/api/v1/end_week', {}, { params });
     // handle http errors first
-    if(response.status !== 200)
-    {
+    if (response.status !== 200) {
       showAlert("Failed to process week end", 'error');
       return;
     }
 
 
     const result = response.data.result;
-    if(result.error)
-    {
+    if (result.error) {
       showAlert(result.error.message, 'error');
     }
-    else if(result.ok)
-    {
+    else if (result.ok) {
       onWeekEnd(user.id);
     }
   }
@@ -99,39 +77,39 @@ const DashboardUserStats: React.FC<DashboardUserStatsProps> = ({
         </div>
 
         <StatBox icon="🎮"
-                 label="Game"
-                 initialValue={gameTimeDisplay}
-                 defaultValue="00:00"
-                 renderPicker={TimeModifier}
-                 applyOperation={(val, op, amount) => update_reward_time(user.id, 'tv_time', val, op, amount)}
-                 renderContent={({ value }) => (
-                   <div className="flex flex-col items-center">
-                     <div className="font-bold text-white text-lg">{value}</div>
-                   </div>)}
+          label="Game"
+          initialValue={gameTimeDisplay}
+          defaultValue="00:00"
+          renderPicker={TimeModifier}
+          applyOperation={(val, op, amount) => update_reward_time(user.id, 'game_time', val, op, amount)}
+          renderContent={({ value }) => (
+            <div className="flex flex-col items-center">
+              <div className="font-bold text-white text-lg">{value}</div>
+            </div>)}
         />
 
         <StatBox icon="📺"
-                 label="TV"
-                 initialValue={tvTimeDisplay}
-                 defaultValue="00:00"
-                 renderPicker={TimeModifier}
-                 applyOperation={(val, op, amount) => update_reward_time(user.id, 'tv_time', val, op, amount)}
-                 renderContent={({ value }) => (
-                   <div className="flex flex-col items-center">
-                     <div className="font-bold text-white text-lg">{value}</div>
-                   </div>)}
+          label="TV"
+          initialValue={tvTimeDisplay}
+          defaultValue="00:00"
+          renderPicker={TimeModifier}
+          applyOperation={(val, op, amount) => update_reward_time(user.id, 'tv_time', val, op, amount)}
+          renderContent={({ value }) => (
+            <div className="flex flex-col items-center">
+              <div className="font-bold text-white text-lg">{value}</div>
+            </div>)}
         />
 
         <StatBox icon="🌟"
-                 label="Stars"
-                 initialValue={stars}
-                 defaultValue={0}
-                 renderPicker={PositiveIntegerPicker}
-                 applyOperation={update_star_points}
-                 renderContent={({ value }) => (
-                   <div className="flex flex-col items-center">
-                     <div className="font-bold text-white text-lg">{value}</div>
-                   </div>)}
+          label="Stars"
+          initialValue={stars}
+          defaultValue={0}
+          renderPicker={PositiveIntegerPicker}
+          applyOperation={update_star_points}
+          renderContent={({ value }) => (
+            <div className="flex flex-col items-center">
+              <div className="font-bold text-white text-lg">{value}</div>
+            </div>)}
         />
 
         <div className="bg-white bg-opacity-20 rounded-xl p-2 text-white">
